+++ java/toolgood.algorithm/src/test/java/toolgood/algorithm/Tests/Cylinder.java
@@ -37,7 +37,7 @@ public class Cylinder extends AlgorithmEngine {
         {
             if (operands.size() == 1)
             {
-                int r =(int) operands.get(0).ToNumber(null).NumberValue();
+                int r =(int) operands.get(0).ToNumber(null).NumberValue().intValue();
                 return Operand.Create(r * r * Math.PI);
             }
         }
